More regular comparison of sizes
With the previous comparison relation of sizes, the successor of
infinity was not less than any other size, not even its own successor
(and similar for the successor of star). This implied that these sizes
were more or less useless. I don't think that this mattered in practice,
but it was a weird corner case.

This commit changes the size comparison relation to be more regular. In
particular, any size is now less than its successor. Also, if n < m then
the successor of n is less than the successor of m -- previously, this
did not hold for all sizes.

This change requires an adjustment to the model of sizes. Previously,
sizes were modelled by the natural numbers or omega. Now we use a larger
set of ordinals.